+++ src/clock/time.go
@@ -1,0 +1,7 @@
+package clock
+
+import "time"
+
+type Time interface {
+	Now() time.Time
+}
