--- src/node/protocol/clock/time.go
@@ -1,13 +1,0 @@
-package clock
-
-import "time"
-
-type Time struct{}
-
-func NewTime() *Time {
-	return &Time{}
-}
-
-func (watch *Time) Now() time.Time {
-	return time.Now()
-}
