+++ src/node/protocol/validation/transactions_pool.go
@@ -3,8 +3,8 @@ package validation
 import (
 	"errors"
 	"fmt"
+	"github.com/my-cloud/ruthenium/src/clock"
 	"github.com/my-cloud/ruthenium/src/log"
-	"github.com/my-cloud/ruthenium/src/node/clock"
 	"github.com/my-cloud/ruthenium/src/node/network"
 	"github.com/my-cloud/ruthenium/src/node/protocol"
 	"math/rand"
